requirements.txt
# Core
Django==5.0.4
gunicorn==22.0.0

# Database
psycopg2-binary==2.9.9

# Environment
python-decouple==3.8

# Email
resend==2.3.0

# Admin UI
django-unfold==0.36.0

# Image handling
Pillow==10.3.0

# Security
django-cors-headers==4.3.1
whitenoise==6.6.0

# Utils
django-extensions==3.2.3
